Added basic filtering for nodes
