+++ raphtory-graphql/src/model/algorithm.rs
@@ -1,4 +1,3 @@
-use crate::model::DynamicGraph;
 use async_graphql::dynamic::{
     Field, FieldFuture, FieldValue, InputValue, Object, ResolverContext, TypeRef,
 };
@@ -11,6 +10,7 @@ use raphtory::db::view_api::GraphViewOps;
 use std::borrow::Cow;
 use std::collections::HashMap;
 use std::sync::Mutex;
+use crate::model::graph::graph::DynamicGraph;
 
 type RegisterFunction = fn(&str, Registry, Object) -> (Registry, Object);
