+++ raphtory-graphql/src/model/filters/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod nodefilter;
