+++ raphtory-graphql/src/model/graph/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod graph;
+pub(crate) mod property;
+pub(crate) mod node;
+pub(crate) mod edge;
